visa_agent/requirements.txt
Django
djangorestframework
crewai
crewai-tools
pyyaml
markdown
django-filter
googlesearch